feat: implement node and edges dropdown filter in graph
