+++ src/components/Universe/GraphSearch/index.tsx
@@ -1,4 +1,3 @@
-import { useLocation } from 'react-router-dom'
 import styled, { css } from 'styled-components'
 import { Flex } from '~/components/common/Flex'
 import ClearIcon from '~/components/Icons/ClearIcon'
@@ -6,14 +5,11 @@ import SearchIcon from '~/components/Icons/SearchIcon'
 import { useAppStore } from '~/stores/useAppStore'
 import { useGraphStore } from '~/stores/useGraphStore'
 import { colors } from '~/utils'
-import { FollowersFilter } from './FollowersFilter'
 import { GraphFilter } from './GraphFilter'
 
 export const GraphSearch = () => {
   const [setSearchQuery, searchQuery] = useGraphStore((s) => [s.setSearchQuery, s.searchQuery])
   const universeQuestionIsOpen = useAppStore((s) => s.universeQuestionIsOpen)
-  const location = useLocation()
-  const isTweetMindset = location.pathname.includes('/tweet/')
 
   if (universeQuestionIsOpen) {
     return null
@@ -48,7 +44,6 @@ export const GraphSearch = () => {
         </SearchWrapper>
         <GraphFilter />
       </TopBarContainer>
-      {isTweetMindset && <FollowersFilter />}
     </Wrapper>
   )
 }
